Add homepage Template
Per @mnorthuis via ticket https://github.com/department-of-veterans-affairs/va.gov-team/issues/82570

Establishing documented guidance on the design of the VA.gov home page, usage of the template, and clear criteria for what content or links can be added to the home page and where it would go. Documentation includes description of each section on the home page and ownership/approval process.